Add a util-service holding methods needed by views
