+++ mico-admin/src/app/app-list/app-list.component.ts
@@ -27,6 +27,7 @@ import { YesNoDialogComponent } from '../dialogs/yes-no-dialog/yes-no-dialog.com
 import { CreateApplicationComponent } from '../dialogs/create-application/create-application.component';
 import { Router } from '@angular/router';
 import { safeUnsubscribe } from '../util/utils';
+import { UtilServiceService } from '../util/util-service.service';
 
 @Component({
     selector: 'mico-app-list',
@@ -42,6 +43,7 @@ export class AppListComponent implements OnInit {
         private dialog: MatDialog,
         private snackBar: MatSnackBar,
         private router: Router,
+        private utilService: UtilServiceService,
     ) { }
 
     applications: Readonly<ApiObject[]>;
@@ -97,39 +99,9 @@ export class AppListComponent implements OnInit {
     }
 
     /**
-     * create a new application
-     * uses: POST application
+     * opens a dialog to create a new application
      */
     newApplication() {
-        const dialogRef = this.dialog.open(CreateApplicationComponent);
-
-        const subDialog = dialogRef.afterClosed().subscribe(result => {
-
-            // filter empty results (when dialog is aborted)
-            if (!result) {
-                return;
-            }
-
-            // check if returned object is complete
-            for (const property in result.applicationProperties) {
-                if (result.applicationProperties[property] == null) {
-
-                    // show an error message containg the missing field
-                    this.snackBar.open('Missing property: ' + property, 'Ok', {
-                        duration: 8000,
-                    });
-                    return;
-                }
-            }
-
-            const data = result.applicationProperties;
-            data.services = result.services;
-
-            this.apiService.postApplication(data).subscribe(val => {
-                this.router.navigate(['app-detail', val.shortName, val.version]);
-            });
-
-            safeUnsubscribe(subDialog);
-        });
+        this.utilService.createNewApplication();
     }
 }
