+++ mico-admin/src/app/dashboard/dashboard.component.ts
@@ -21,12 +21,11 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { ApiService } from '../api/api.service';
 import { ApiObject } from '../api/apiobject';
 import { MatDialog, MatSnackBar } from '@angular/material';
-import { CreateServiceDialogComponent } from '../dialogs/create-service/create-service.component';
 import { Router } from '@angular/router';
-import { CreateApplicationComponent } from '../dialogs/create-application/create-application.component';
 import { groupBy, mergeMap, toArray, map } from 'rxjs/operators';
 import { Subscription, from } from 'rxjs';
 import { safeUnsubscribe } from '../util/utils';
+import { UtilServiceService } from '../util/util-service.service';
 
 @Component({
     selector: 'mico-dashboard',
@@ -37,9 +36,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
 
     constructor(
         private apiService: ApiService,
-        private dialog: MatDialog,
-        private router: Router,
-        private snackBar: MatSnackBar,
+        private utilService: UtilServiceService,
     ) { }
 
     subApplications: Subscription;
@@ -84,48 +81,10 @@ export class DashboardComponent implements OnInit, OnDestroy {
 
 
     /**
-     * dialog to create a new service. can be done:
-     * - manually
-     * - via github import
-     * uses: POST services or POST services/import/github
+     * opens a dialog to create a new service.
      */
     newService(): void {
-        const dialogRef = this.dialog.open(CreateServiceDialogComponent);
-
-        const subDialog = dialogRef.afterClosed().subscribe(result => {
-
-            // filter empty results (when dialog is aborted)
-            if (!result) {
-                return;
-            }
-
-            // check if returned object is complete
-            for (const property in result.data) {
-                if (result.data[property] == null) {
-
-                    if (property !== 'serviceInterfaces') {
-                        this.snackBar.open('Missing property: ' + property, 'Ok', {
-                            duration: 8000,
-                        });
-                        return;
-                    }
-                }
-            }
-
-            // decide if the service was created manually or is to be created via github crawler and create service
-            if (result.tab === 'manual') {
-                this.apiService.postService(result.data).subscribe(val => {
-                    this.router.navigate(['service-detail', val.shortName, val.version]);
-                });
-            } else if (result.tab === 'github') {
-
-                this.apiService.postServiceViaGithub(result.data.url, result.data.version).subscribe(val => {
-                    this.router.navigate(['service-detail', val.shortName, val.version]);
-                });
-            }
-
-            safeUnsubscribe(subDialog);
-        });
+        this.utilService.createNewService();
     }
 
     /**
@@ -133,36 +92,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
      * uses: POST application
      */
     newApplication() {
-        const dialogRef = this.dialog.open(CreateApplicationComponent);
-
-        const subDialog = dialogRef.afterClosed().subscribe(result => {
-
-            // filter empty results (when dialog is aborted)
-            if (!result) {
-                return;
-            }
-
-            // check if returned object is complete
-            for (const property in result.applicationProperties) {
-                if (result.applicationProperties[property] == null) {
-
-                    // show an error message containg the missing field
-                    this.snackBar.open('Missing property: ' + property, 'Ok', {
-                        duration: 8000,
-                    });
-                    return;
-                }
-            }
-
-            const data = result.applicationProperties;
-            data.services = result.services;
-
-            this.apiService.postApplication(data).subscribe(val => {
-                this.router.navigate(['app-detail', val.shortName, val.version]);
-            });
-
-            safeUnsubscribe(subDialog);
-        });
+        this.utilService.createNewApplication();
     }
 
     // links to our related pages
