+++ mico-admin/src/app/service-list/service-list.component.ts
@@ -28,6 +28,7 @@ import { YesNoDialogComponent } from '../dialogs/yes-no-dialog/yes-no-dialog.com
 import { Router } from '@angular/router';
 import { CreateServiceDialogComponent } from '../dialogs/create-service/create-service.component';
 import { safeUnsubscribe } from '../util/utils';
+import { UtilServiceService } from '../util/util-service.service';
 
 
 @Component({
@@ -42,8 +43,7 @@ export class ServiceListComponent implements OnInit, OnDestroy {
     constructor(
         private apiService: ApiService,
         private dialog: MatDialog,
-        private router: Router,
-        private snackBar: MatSnackBar,
+        private utilService: UtilServiceService,
     ) {
         this.getServices();
     }
@@ -108,47 +108,9 @@ export class ServiceListComponent implements OnInit, OnDestroy {
 
 
     /**
-     * dialog to create a new service. can be done:
-     * - manually
-     * - via github import
-     * uses: POST services or POST services/import/github
+     * opens a dialog to create a new service.
      */
     newService(): void {
-        const dialogRef = this.dialog.open(CreateServiceDialogComponent);
-
-        const subDialog = dialogRef.afterClosed().subscribe(result => {
-
-            // filter empty results (when dialog is aborted)
-            if (!result) {
-                return;
-            }
-
-            // check if returned object is complete
-            for (const property in result.data) {
-                if (result.data[property] == null) {
-
-                    if (property !== 'serviceInterfaces') {
-                        this.snackBar.open('Missing property: ' + property, 'Ok', {
-                            duration: 8000,
-                        });
-                        return;
-                    }
-                }
-            }
-
-            // decide if the service was created manually or is to be created via github crawler and create service
-            if (result.tab === 'manual') {
-                this.apiService.postService(result.data).subscribe(val => {
-                    this.router.navigate(['service-detail', val.shortName, val.version]);
-                });
-            } else if (result.tab === 'github') {
-
-                this.apiService.postServiceViaGithub(result.data.url, result.data.version).subscribe(val => {
-                    this.router.navigate(['service-detail', val.shortName, val.version]);
-                });
-            }
-
-            safeUnsubscribe(subDialog);
-        });
+        this.utilService.createNewService();
     }
 }
